modified and refactored test files
