+++ src/test/resources/TestingResources/TestCases/HotCold.js
